module1-qa-cli/pyproject.toml
[project]
name = "lc-module1-qa-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=26.3.1",
    "isort>=8.0.1",
    "langchain>=1.2.14",
    "langchain-core>=1.2.23",
    "langchain-openai>=1.1.12",
    "python-dotenv>=1.2.2",
]
